ABIChecker: minor refactoring to move code to APIDigester lib. NFC
